pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[virtualenvs]
create = true
in-project = true


[tool.poetry]
name = "hedgedoc-history"
version = "0.1.0"
description = ""
authors = ["gvido-berzins <gvido.berzins@syberu.xyz>"]
readme = "README.md"
packages = [{include = "hedgedoc_history"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
loguru = "^0.6.0"
cattrs = "^22.2.0"
attrs = "^22.1.0"
attrs-strict = "^1.0.0"
pyyaml = "^6.0"
environs = "^9.5.0"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"

[tool.poetry.scripts]
hdd = "hedgedoc_history.cli:main"


[tool.black]
line-length = 90

[tool.isort]
line_length = 90
profile = "black"

[tool.flake8]
max-line-length = 90
